Move shared chains sdk to common
