+++ cmd/main.go
@@ -24,6 +24,7 @@ import (
 	"os/signal"
 	"poly-bridge/chainfeelisten"
 	"poly-bridge/coinpricelisten"
+	"poly-bridge/common"
 	"poly-bridge/conf"
 	"poly-bridge/crosschaineffect"
 	"poly-bridge/crosschainlisten"
@@ -109,6 +110,7 @@ func startServer(ctx *cli.Context) {
 		conf, _ := json.Marshal(config)
 		logs.Info("%s\n", string(conf))
 	}
+	common.SetupChainsSDK(config)
 	crosschainlisten.StartCrossChainListen(config.Server, config.Backup, config.ChainListenConfig, config.DBConfig)
 	if config.Backup {
 		return
