+++ main.go
@@ -19,15 +19,24 @@ package main
 
 import (
 	"encoding/json"
+	"poly-bridge/common"
+	"poly-bridge/conf"
+	_ "poly-bridge/routers"
+
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/context"
 	"github.com/astaxie/beego/logs"
 	"github.com/astaxie/beego/plugins/cors"
-	_ "poly-bridge/routers"
 )
 
 func main() {
+
+	// Initialize
 	logs.SetLogger(logs.AdapterFile, `{"filename":"logs/bridge_http.log"}`)
+	configFile := beego.AppConfig.String("chain_config")
+	config := conf.NewConfig(configFile)
+	common.SetupChainsSDK(config)
+
 	mode := beego.AppConfig.String("runmode")
 	if mode == "dev" {
 		var FilterLog = func(ctx *context.Context) {
